impl: move the cache management to caches.py
Separate cache management from main algorithms, update docs and tests
